finish partitioning
