+++ hw3/animal_management/animal_manager.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import Any, List, Optional
 
 from hw3.animal_management.animal import Animal
 
@@ -15,4 +15,10 @@ class AnimalManager:
         pass
 
     def remove_animal(self, animal_id: int) -> None:
+        pass
+
+    def get_animal_details(animal_id) -> dict[str, Any]:
+        pass
+
+    def get_animals_in_habitat(habitat_id: int) -> List[Animal]:
         pass
