+++ hw3/formal_specification.py
@@ -31,16 +31,16 @@ start_location: Habitat
 status: str = "Scheduled"
 
 
-def assign_animals_to_habitat(animals: List[Animal]) -> None:
-    pass
+# def assign_animals_to_habitat(animals: List[Animal]) -> None:
+#     pass
 
 
-def assign_animals_to_habitat(habitat_id: int, animals: List[Animal]) -> None:
-    pass
+# def assign_animals_to_habitat(habitat_id: int, animals: List[Animal]) -> None:
+#     pass
 
 
-def cancel_migration(migration_id: int) -> None:
-    pass
+# def cancel_migration(migration_id: int) -> None:
+#     pass
 
 
 # def create_habitat(
@@ -49,13 +49,13 @@ def cancel_migration(migration_id: int) -> None:
 #     pass
 
 
-def create_migration_path(
-    species: str,
-    start_location: Habitat,
-    destination: Habitat,
-    duration: Optional[int] = None,
-) -> None:
-    pass
+# def create_migration_path(
+#     species: str,
+#     start_location: Habitat,
+#     destination: Habitat,
+#     duration: Optional[int] = None,
+# ) -> None:
+#     pass
 
 
 # def get_animal_by_id(animal_id: int) -> Optional[Animal]:
@@ -66,82 +66,82 @@ def create_migration_path(
 #     pass
 
 
-def get_animals_in_habitat(habitat_id: int) -> List[Animal]:
-    pass
+# def get_animals_in_habitat(habitat_id: int) -> List[Animal]:
+#     pass
 
 
 # def get_habitat_by_id(habitat_id: int) -> Habitat:
 #     pass
 
 
-def get_habitat_details(habitat_id: int) -> dict:
-    pass
+# def get_habitat_details(habitat_id: int) -> dict:
+#     pass
 
 
-def get_habitats_by_geographic_area(geographic_area: str) -> List[Habitat]:
-    pass
+# def get_habitats_by_geographic_area(geographic_area: str) -> List[Habitat]:
+#     pass
 
 
-def get_habitats_by_size(size: int) -> List[Habitat]:
-    pass
+# def get_habitats_by_size(size: int) -> List[Habitat]:
+#     pass
 
 
-def get_habitats_by_type(environment_type: str) -> List[Habitat]:
-    pass
+# def get_habitats_by_type(environment_type: str) -> List[Habitat]:
+#     pass
 
 
-def get_migration_by_id(migration_id: int) -> Migration:
-    pass
+# def get_migration_by_id(migration_id: int) -> Migration:
+#     pass
 
 
-def get_migration_details(migration_id: int) -> dict[str, Any]:
-    pass
+# def get_migration_details(migration_id: int) -> dict[str, Any]:
+#     pass
 
 
-def get_migration_path_by_id(path_id: int) -> MigrationPath:
-    pass
+# def get_migration_path_by_id(path_id: int) -> MigrationPath:
+#     pass
 
 
-def get_migration_paths() -> list[MigrationPath]:
-    pass
+# def get_migration_paths() -> list[MigrationPath]:
+#     pass
 
 
-def get_migration_paths_by_destination(destination: Habitat) -> list[MigrationPath]:
-    pass
+# def get_migration_paths_by_destination(destination: Habitat) -> list[MigrationPath]:
+#     pass
 
 
-def get_migration_paths_by_species(species: str) -> list[MigrationPath]:
-    pass
+# def get_migration_paths_by_species(species: str) -> list[MigrationPath]:
+#     pass
 
 
-def get_migration_paths_by_start_location(
-    start_location: Habitat,
-) -> list[MigrationPath]:
-    pass
+# def get_migration_paths_by_start_location(
+#     start_location: Habitat,
+# ) -> list[MigrationPath]:
+#     pass
 
 
-def get_migrations() -> list[Migration]:
-    pass
+# def get_migrations() -> list[Migration]:
+#     pass
 
 
-def get_migrations_by_current_location(current_location: str) -> list[Migration]:
-    pass
+# def get_migrations_by_current_location(current_location: str) -> list[Migration]:
+#     pass
 
 
-def get_migrations_by_migration_path(migration_path_id: int) -> list[Migration]:
-    pass
+# def get_migrations_by_migration_path(migration_path_id: int) -> list[Migration]:
+#     pass
 
 
-def get_migrations_by_start_date(start_date: str) -> list[Migration]:
-    pass
+# def get_migrations_by_start_date(start_date: str) -> list[Migration]:
+#     pass
 
 
-def get_migrations_by_status(status: str) -> list[Migration]:
-    pass
+# def get_migrations_by_status(status: str) -> list[Migration]:
+#     pass
 
 
-def get_migration_path_details(path_id) -> dict:
-    pass
+# def get_migration_path_details(path_id) -> dict:
+#     pass
 
 
 # def register_animal(animal: Animal) -> None:
@@ -156,25 +156,25 @@ def get_migration_path_details(path_id) -> dict:
 #     pass
 
 
-def remove_migration_path(path_id: int) -> None:
-    pass
+# def remove_migration_path(path_id: int) -> None:
+#     pass
 
 
-def schedule_migration(migration_path: MigrationPath) -> None:
-    pass
+# def schedule_migration(migration_path: MigrationPath) -> None:
+#     pass
 
 
 # def update_animal_details(animal_id: int, **kwargs: Any) -> None:
 #     pass
 
 
-def update_habitat_details(habitat_id: int, **kwargs: dict[str, Any]) -> None:
-    pass
+# def update_habitat_details(habitat_id: int, **kwargs: dict[str, Any]) -> None:
+#     pass
 
 
-def update_migration_details(migration_id: int, **kwargs: Any) -> None:
-    pass
+# def update_migration_details(migration_id: int, **kwargs: Any) -> None:
+#     pass
 
 
-def update_migration_path_details(path_id: int, **kwargs) -> None:
-    pass
+# def update_migration_path_details(path_id: int, **kwargs) -> None:
+#     pass
