+++ hw3/migration_tracking/migration_path.py
@@ -1,5 +1,36 @@
 from typing import Optional
 
+from hw3.habitat_management.habitat import Habitat
+from hw3.migration_tracking.migration_path import MigrationPath
+
+
 class MigrationPath:
 
-    pass
+    def __init__(
+        self,
+        path_id: int,
+        start_location: Habitat,
+        destination: Habitat,
+        duration: Optional[int] = None,
+    ):
+        self.path_id = path_id
+        self.start_location = start_location
+        self.destination = destination
+        self.duration = duration
+
+    def get_migration_paths() -> list[MigrationPath]:
+        pass
+
+    def get_migration_paths_by_destination(destination: Habitat) -> list[MigrationPath]:
+        pass
+
+    def get_migration_paths_by_species(species: str) -> list[MigrationPath]:
+        pass
+
+    def get_migration_paths_by_start_location(
+        start_location: Habitat,
+    ) -> list[MigrationPath]:
+        pass
+
+    def update_migration_path_details(path_id: int, **kwargs) -> None:
+        pass
